Fix OTP verification and registration button issues
- Fix OTP code type mismatch: cast to (int) everywhere (store + verify)
- Fix mailVerificationTemplate: cast code to (int)
- Fix register page: Personal Account selected by default, form visible
- Fix OTP input: type='text' with inputmode='numeric' for mobile visibility
- Fix mailVerify redirect: preserve token on error
- Full OTP + email_verified update tested end-to-end

+++ app/Http/Helpers/helpers.php
@@ -1489,7 +1489,7 @@ function mailVerificationTemplate($user)
 {
     $data = [
         'user_id' => $user->id,
-        'code' => generate_random_code(),
+        'code' => (int) generate_random_code(),
         'token' => generate_unique_string('user_authorizations', 'token', 200),
         'created_at' => now(),
     ];
@@ -1891,7 +1891,7 @@ function mailVerificationTemplateApi($user)
 
     $data = [
         'user_id' => $user->id,
-        'code' => generate_random_code(),
+        'code' => (int) generate_random_code(),
         'token' => generate_unique_string('user_authorizations', 'token', 200),
         'created_at' => now(),
     ];
